PIM-5714: Move reference data serializers in ReferenceData component
