+++ src/Pim/Bundle/CatalogBundle/Resources/config/serializers.yml
@@ -23,7 +23,6 @@ parameters:
     pim_serializer.normalizer.product_price.class:          Pim\Component\Catalog\Normalizer\Structured\ProductPriceNormalizer
     pim_serializer.normalizer.product_values.class:         Pim\Component\Catalog\Normalizer\Structured\ProductValuesNormalizer
     pim_serializer.normalizer.product_value.class:          Pim\Component\Catalog\Normalizer\Structured\ProductValueNormalizer
-    pim_serializer.normalizer.reference_data.class:         Pim\Component\Catalog\Normalizer\Structured\ReferenceDataNormalizer
     pim_serializer.normalizer.file.class:                   Pim\Component\Catalog\Normalizer\Structured\FileNormalizer
 
     # Denormalizers
@@ -38,8 +37,6 @@ parameters:
     pim_serializer.denormalizer.datetime.class:                  Pim\Component\Catalog\Denormalizer\Structured\ProductValue\DateTimeDenormalizer
     pim_serializer.denormalizer.file.class:                      Pim\Component\Catalog\Denormalizer\Structured\ProductValue\FileDenormalizer
     pim_serializer.denormalizer.boolean.class:                   Pim\Component\Catalog\Denormalizer\Structured\ProductValue\BooleanDenormalizer
-    pim_serializer.denormalizer.reference_data.class:            Pim\Component\Catalog\Denormalizer\Structured\ProductValue\ReferenceDataDenormalizer
-    pim_serializer.denormalizer.reference_data_collection.class: Pim\Component\Catalog\Denormalizer\Structured\ProductValue\ReferenceDataCollectionDenormalizer
 
 services:
     # Serializer, TODO : should be moved somewhere else
@@ -76,11 +73,6 @@ services:
         class: %pim_serializer.normalizer.product_value.class%
         arguments:
             - '@pim_catalog.localization.localizer.registry'
-        tags:
-            - { name: pim_serializer.normalizer, priority: 90 }
-
-    pim_serializer.normalizer.reference_data:
-        class: %pim_serializer.normalizer.reference_data.class%
         tags:
             - { name: pim_serializer.normalizer, priority: 90 }
 
@@ -256,21 +248,5 @@ services:
         class: %pim_serializer.denormalizer.boolean.class%
         arguments:
             - ['pim_catalog_boolean']
-        tags:
-            - { name: pim_serializer.normalizer, priority: 90 }
-
-    pim_serialize.denormalizer.reference_data:
-        class: %pim_serializer.denormalizer.reference_data.class%
-        arguments:
-            - ['pim_reference_data_simpleselect']
-            - '@?pim_reference_data.repository_resolver'
-        tags:
-            - { name: pim_serializer.normalizer, priority: 90 }
-
-    pim_serialize.denormalizer.reference_data_collection:
-        class: %pim_serializer.denormalizer.reference_data_collection.class%
-        arguments:
-            - ['pim_reference_data_multiselect']
-            - '@pim_serialize.denormalizer.reference_data'
         tags:
             - { name: pim_serializer.normalizer, priority: 90 }
