+++ src/Pim/Bundle/ConnectorBundle/Resources/config/serializer.yml
@@ -21,7 +21,6 @@ parameters:
     pim_serializer.normalizer.flat.metric.class:            Pim\Component\Connector\Normalizer\Flat\MetricNormalizer
     pim_serializer.normalizer.flat.option.class:            Pim\Component\Connector\Normalizer\Flat\AttributeOptionNormalizer
     pim_serializer.normalizer.flat.price.class:             Pim\Component\Connector\Normalizer\Flat\PriceNormalizer
-    pim_serializer.normalizer.flat.reference_data.class:    Pim\Component\Connector\Normalizer\Flat\ReferenceDataNormalizer
     pim_serializer.normalizer.flat.file.class:              Pim\Component\Connector\Normalizer\Flat\FileNormalizer
 
      # Denormalizers
@@ -39,8 +38,6 @@ parameters:
     pim_serializer.denormalizer.flat.metric.class:                    Pim\Component\Connector\Denormalizer\Flat\ProductValue\MetricDenormalizer
     pim_serializer.denormalizer.flat.datetime.class:                  Pim\Component\Connector\Denormalizer\Flat\ProductValue\DateTimeDenormalizer
     pim_serializer.denormalizer.flat.file.class:                      Pim\Component\Connector\Denormalizer\Flat\ProductValue\FileDenormalizer
-    pim_serializer.denormalizer.flat.reference_data.class:            Pim\Component\Connector\Denormalizer\Flat\ProductValue\ReferenceDataDenormalizer
-    pim_serializer.denormalizer.flat.reference_data_collection.class: Pim\Component\Connector\Denormalizer\Flat\ProductValue\ReferenceDataCollectionDenormalizer
 
 services:
     # Encoder
@@ -152,11 +149,6 @@ services:
 
     pim_serializer.normalizer.flat.price:
         class: %pim_serializer.normalizer.flat.price.class%
-        tags:
-            - { name: pim_serializer.normalizer, priority: 90 }
-
-    pim_serializer.normalizer.flat.reference_data:
-        class: %pim_serializer.normalizer.flat.reference_data.class%
         tags:
             - { name: pim_serializer.normalizer, priority: 90 }
 
@@ -281,22 +273,6 @@ services:
             - ['pim_catalog_image', 'pim_catalog_file']
             - '@akeneo_file_storage.repository.file_info'
             - '@akeneo_file_storage.file_storage.file.file_storer'
-        tags:
-            - { name: pim_serializer.normalizer, priority: 90 }
-
-    pim_serializer.denormalizer.flat.reference_data:
-        class: %pim_serializer.denormalizer.flat.reference_data.class%
-        arguments:
-            - ['pim_reference_data_simpleselect']
-            - '@?pim_reference_data.repository_resolver'
-        tags:
-            - { name: pim_serializer.normalizer, priority: 90 }
-
-    pim_serializer.denormalizer.flat.reference_data_collection:
-        class: %pim_serializer.denormalizer.flat.reference_data_collection.class%
-        arguments:
-            - ['pim_reference_data_multiselect']
-            - '@?pim_reference_data.repository_resolver'
         tags:
             - { name: pim_serializer.normalizer, priority: 90 }
