+++ src/Pim/Bundle/ReferenceDataBundle/Resources/config/serializers.yml
@@ -1,0 +1,61 @@
+parameters:
+    # Normalizers
+    pim_reference_data.normalizer.configuration.class:       Pim\Bundle\ReferenceDataBundle\Normalizer\ConfigurationNormalizer
+    pim_reference_data.normalizer.reference_data.class:      Pim\Component\ReferenceData\Normalizer\Structured\ReferenceDataNormalizer
+    pim_reference_data.normalizer.flat.reference_data.class: Pim\Component\ReferenceData\Normalizer\Flat\ReferenceDataNormalizer
+
+    # Denormalizers
+    pim_reference_data.denormalizer.reference_data.class:                 Pim\Component\ReferenceData\Denormalizer\Structured\ProductValue\ReferenceDataDenormalizer
+    pim_reference_data.denormalizer.reference_data_collection.class:      Pim\Component\ReferenceData\Denormalizer\Structured\ProductValue\ReferenceDataCollectionDenormalizer
+    pim_reference_data.denormalizer.flat.reference_data.class:            Pim\Component\ReferenceData\Denormalizer\Flat\ProductValue\ReferenceDataDenormalizer
+    pim_reference_data.denormalizer.flat.reference_data_collection.class: Pim\Component\ReferenceData\Denormalizer\Flat\ProductValue\ReferenceDataCollectionDenormalizer
+
+services:
+    # Normalizers
+    pim_reference_data.normalizer.configuration:
+        class: '%pim_reference_data.normalizer.configuration.class%'
+        tags:
+            - { name: pim_internal_api_serializer.normalizer }
+
+    pim_reference_data.normalizer.reference_data:
+        class: '%pim_reference_data.normalizer.reference_data.class%'
+        tags:
+            - { name: pim_serializer.normalizer, priority: 90 }
+
+    pim_reference_data.normalizer.flat.reference_data:
+        class: '%pim_reference_data.normalizer.flat.reference_data.class%'
+        tags:
+            - { name: pim_serializer.normalizer, priority: 90 }
+
+    # Denormalizers
+    pim_reference_data.denormalizer.reference_data:
+        class: '%pim_reference_data.denormalizer.reference_data.class%'
+        arguments:
+            - ['pim_reference_data_simpleselect']
+            - '@?pim_reference_data.repository_resolver'
+        tags:
+            - { name: pim_serializer.normalizer, priority: 90 }
+
+    pim_reference_data.denormalizer.reference_data_collection:
+        class: '%pim_reference_data.denormalizer.reference_data_collection.class%'
+        arguments:
+            - ['pim_reference_data_multiselect']
+            - '@pim_reference_data.denormalizer.reference_data'
+        tags:
+            - { name: pim_serializer.normalizer, priority: 90 }
+
+    pim_reference_data.denormalizer.flat.reference_data:
+        class: '%pim_reference_data.denormalizer.flat.reference_data.class%'
+        arguments:
+            - ['pim_reference_data_simpleselect']
+            - '@?pim_reference_data.repository_resolver'
+        tags:
+            - { name: pim_serializer.normalizer, priority: 90 }
+
+    pim_reference_data.denormalizer.flat.reference_data_collection:
+        class: '%pim_reference_data.denormalizer.flat.reference_data_collection.class%'
+        arguments:
+            - ['pim_reference_data_multiselect']
+            - '@?pim_reference_data.repository_resolver'
+        tags:
+            - { name: pim_serializer.normalizer, priority: 90 }
