+++ src/Pim/Bundle/VersioningBundle/Resources/config/serializer/flat.yml
@@ -102,7 +102,7 @@ services:
             - { name: pim_versioning.serializer.normalizer, priority: 90 }
 
     pim_versioning.serializer.normalizer.flat.reference_data:
-        class: %pim_serializer.normalizer.flat.reference_data.class%
+        class: %pim_reference_data.normalizer.flat.reference_data.class%
         tags:
             - { name: pim_versioning.serializer.normalizer, priority: 90 }
