+++ src/Pim/Bundle/WebServiceBundle/Resources/config/serializer/structured.yml
@@ -40,6 +40,6 @@ services:
             - { name: pim_webservice.serializer.normalizer, priority: 90 }
 
     pim_webservice.serializer.normalizer.reference_data:
-        class: %pim_serializer.normalizer.reference_data.class%
+        class: %pim_reference_data.normalizer.reference_data.class%
         tags:
             - { name: pim_webservice.serializer.normalizer, priority: 90 }
